CLI 'forget' to remove *deployments.txt
Enables a subset of the `nile clean` command.

Workflow loop of (edit-compile-deploy-interact) is hampered by an error with alias name being occupied by the previous loop. 

```
Exception: Alias MyContract already exists in localhost.deployments.txt
```

This command adds the ability to erase the deployments file for the current network. This feature is provided in `nile clean`, however 
in a project with many contracts `nile clean` removing the compiled contracts directory is an issue. New loop is: (edit-compile-forget-deploy-interact)